Ajout validation + vérification signature
